% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_hyper_init.R
\name{feed_hyperparam}
\alias{feed_hyperparam}
\title{Gather model hyperparameters provided by the user.}
\usage{
feed_hyperparam(d, p, eta, kappa, lambda, nu, a, b, q = NULL, phi = NULL,
  xi = NULL)
}
\description{
Gather model hyperparameters provided by the user.
}

