% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_null.R
\name{generate_null}
\alias{generate_null}
\title{Generate empirical null distribution using permuted data.}
\usage{
generate_null(n_perm, Y, X, p0_av, Z, list_hyper, list_init, list_blocks,
  user_seed, tol, maxit, batch, verbose, results_dir, n_cpus)
}
\description{
Generate empirical null distribution using permuted data.
}

