% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_snps}
\alias{generate_snps}
\title{Generate SNP data with prespecified spatial correlation structure.}
\usage{
generate_snps(n, p, cor_type = NULL, vec_rho = NULL, vec_maf = NULL,
  n_cpus = 1)
}
\description{
Generate SNP data with prespecified spatial correlation structure.
}

