% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horseshoe_core.R
\name{horseshoe_core}
\alias{horseshoe_core}
\title{Horseshoe sparse regression with Cauchy priors on variances}
\usage{
horseshoe_core(Y, X, d, n, p, list_hyper, alpha_vb, mu_beta_vb, sig2_beta_vb,
  tau_vb, tol, maxit, verbose, shared_prec, full_output = FALSE)
}
