% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horseshoe_core_exp.R
\name{horseshoe_core_exp}
\alias{horseshoe_core_exp}
\title{Horseshoe sparse regression with exponential family reduction and Cauchy priors on variances}
\usage{
horseshoe_core_exp(Y, X, d, n, p, list_hyper, alpha_vb, c_vb, mu_beta_vb,
  sig2_beta_vb, tau_vb, tol, maxit, verbose, shared_prec = F,
  full_output = FALSE)
}
