% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horseshoe_core_gamma.R
\name{horseshoe_core_gamma}
\alias{horseshoe_core_gamma}
\title{horseshoe sparse regression with gamma priors on variances}
\usage{
horseshoe_core_gamma(Y, X, d, n, p, list_hyper, alpha_vb, mu_beta_vb,
  sig2_beta_vb, tau_vb, tol, maxit, verbose, shared_prec = T,
  full_output = FALSE)
}
