% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horseshoe_core_plus.R
\name{horseshoe_core_plus}
\alias{horseshoe_core_plus}
\title{Horseshoe + sparse regression with Cauchy priors on variances parameters}
\usage{
horseshoe_core_plus(Y, X, d, n, p, list_hyper, alpha_vb, c_vb, d_vb, e_vb,
  mu_beta_vb, sig2_beta_vb, tau_vb, tol, maxit, batch, verbose,
  shared_prec = F, full_output = FALSE)
}
