% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus-package.R
\docType{package}
\name{locus-package}
\alias{locus}
\alias{locus-package}
\title{locus: Location-based Datasets for the U.S.}
\description{
Location-based datasets for the U.S. related
    to population, quality of life, social determinants of health, and
    climate susceptibility
}
\author{
\strong{Maintainer}: Brian Vancil \email{brian@vancil.name} (\href{https://orcid.org/https://orcid.org/0000-0003-1145-3681}{ORCID})

}
\keyword{internal}
