% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_horseshoe_core.R
\name{locus_core_horseshoe}
\alias{locus_core_horseshoe}
\title{HS locus}
\usage{
locus_core_horseshoe(Y, X, d, n, p, list_hyper, b_vb, sigma2_bv, mu_beta_vb,
  sig2_beta_vb, tau_vb, tol, maxit, batch, verbose, scheme = "noPrec",
  loop = "c++", full_output = FALSE)
}
\description{
HS locus
}

