% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_horseshoeExp_core.R
\name{locus_core_horseshoeExp}
\alias{locus_core_horseshoeExp}
\title{HS locus Exp}
\usage{
locus_core_horseshoeExp(Y, X, d, n, p, list_hyper, b_vb, c_vb, sigma2_vb,
  mu_beta_vb, sig2_beta_vb, tau_vb, tol, maxit, batch, verbose,
  scheme = "noPrec", loop = "c++", full_output = FALSE)
}
\description{
HS locus Exp
}

