% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{replicate_real_phenos}
\alias{replicate_real_phenos}
\title{Generate phenotypes emulating real phenotypic data at hand.}
\usage{
replicate_real_phenos(n, real_phenos, bl_lgth = NULL, d = NULL,
  n_cpus = 1)
}
\description{
Generate phenotypes emulating real phenotypic data at hand.
}

