% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{replicate_real_snps}
\alias{replicate_real_snps}
\title{Generate SNPs emulating real SNP data at hand.}
\usage{
replicate_real_snps(n, real_snps, bl_lgth, p = NULL, maf_thres = NULL,
  n_cpus = 1)
}
\description{
Generate SNPs emulating real SNP data at hand.
}

