% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_locus.R
\name{set_blocks}
\alias{set_blocks}
\title{Gather settings for parallel inference on partitioned predictor space.}
\usage{
set_blocks(p, pos_bl, n_cpus, verbose = T)
}
\description{
Gather settings for parallel inference on partitioned predictor space.
}

