% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validate.R
\name{set_cv}
\alias{set_cv}
\title{Gather settings for the cross-validation procedure used in \code{locus}.}
\usage{
set_cv(n, p, n_folds, size_p0_av_grid, n_cpus, tol_cv, maxit_cv = 1000,
  batch_cv = T, verbose = T)
}
\description{
Gather settings for the cross-validation procedure used in \code{locus}.
}

