% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_hyper_init.R
\name{feed_init_param}
\alias{feed_init_param}
\title{Gather initial variational parameters provided by the user.}
\usage{
feed_init_param(d, p, gam_vb, mu_beta_vb, sig2_beta_vb, tau_vb, q = NULL,
  mu_alpha_vb = NULL, sig2_alpha_vb = NULL, zeta2_inv_vb = NULL)
}
\description{
Gather initial variational parameters provided by the user.
}

