% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_dependence}
\alias{generate_dependence}
\title{Generate pleiotropic associations between phenotypes and SNPs.}
\usage{
generate_dependence(list_snps, list_phenos, ind_d0, ind_p0, vec_prob_sh,
  pve_per_snp = NULL, max_tot_pve = NULL)
}
\description{
Generate pleiotropic associations between phenotypes and SNPs.
}

