% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_null.R
\name{generate_null}
\alias{generate_null}
\title{Generate empirical null distribution using permuted data.}
\usage{
generate_null(n_perm, Y, X, p0_av, Z = NULL, link = "identity",
  ind_bin = NULL, list_hyper = NULL, list_init = NULL,
  list_blocks = NULL, list_groups = NULL, list_struct = NULL,
  user_seed = NULL, tol = 0.001, maxit = 1000, verbose = TRUE,
  results_dir = NULL, n_cpus = 1)
}
\arguments{
\item{n_perm}{Number of permuted datasets on which \code{\link{locus}} is
run.}

\item{Y}{Response data matrix (without permuted sample indices) of size
n x d, where n is the number of samples and d is the number of
response variables.}

\item{X}{Input matrix of size n x p, where p is the number of candidate
predictors. \code{X} cannot contain NAs. No intercept must be supplied.}

\item{p0_av}{Prior average number of predictors expected to be included in
the model (if \code{list_blocks} is \code{TRUE}, average number per
predictor blocks). Must be \code{NULL} if \code{list_init} and
\code{list_hyper} are both non-\code{NULL} or if \code{list_cv} is
non-\code{NULL}. If \code{list_blocks} is \code{NULL}, can also be a vector
of length p with entry s corresponding to the prior probability that
candidate predictor s is associated with at least one response.}

\item{Z}{Covariate matrix of size n x q, where q is the number of
covariates. \code{NULL} if no covariate. Factor covariates must be supplied
after transformation to dummy coding. No intercept must be supplied.}

\item{link}{Response link. Must be "\code{identity}" for linear regression,
"\code{logit}" for logistic regression, "\code{probit}" for probit
regression, or "\code{mix}" for a mix of identity and probit link functions
(in this case, the indices of the binary responses must be gathered in
argument \code{ind_bin}, see below).}

\item{ind_bin}{If \code{link = "mix"}, vector of indices corresponding to
the binary variables in \code{Y}. Must be \code{NULL} if
\code{link != "mix"}.}

\item{list_hyper}{An object of class "\code{hyper}" containing the model
hyperparameters. Must be filled using the \code{\link{set_hyper}}
function or must be \code{NULL} for default hyperparameters.}

\item{list_init}{An object of class "\code{init}" containing the initial
variational parameters. Must be filled using the \code{\link{set_init}}
function or be \code{NULL} for a default initialization.}

\item{list_blocks}{An object of class "\code{blocks}" containing setting for
parallel inference on a partitioned predictor space. Must be filled using
the \code{\link{set_blocks}} function or must be \code{NULL} for no
partitioning.}

\item{list_groups}{An object of class "\code{groups}" containing settings for
group selection of candidate predictors. Must be filled using the
\code{\link{set_groups}} function or must be \code{NULL} for group
selection.}

\item{list_struct}{An object of class "\code{struct}" containing settings for
structure sparsity priors. Must be filled using the
\code{\link{set_struct}} function or must be \code{NULL} for structured
selection.}

\item{user_seed}{Seed set for reproducible default choices of hyperparameters
(if \code{list_hyper} is \code{NULL}) and inital variational parameters (if
\code{list_init} is \code{NULL}). Default is \code{NULL}, no seed set.}

\item{tol}{Tolerance for the stopping criterion.}

\item{maxit}{Maximum number of iterations allowed.}

\item{verbose}{If \code{TRUE}, messages are displayed during execution.}

\item{results_dir}{Path where the output of each of the \code{n_perm} runs
will be saved. Default is \code{NULL}, the output is not saved to files and
a list of objects of class "\code{perm}" is returned instead.}

\item{n_cpus}{Number of CPUs to be used. If large, one should ensure that
enough RAM will be available for parallel execution. Set to 1 for serial
execution.}
}
\value{
If \code{results_dir} is \code{NULL}, list of length \code{n_perm}
  with each element corresponding to the output of \code{\link{locus}} on one
  permuted dataset. Each contains:
 \item{ind_perm}{Vector of length n containing the permuted response sample
                 labels.}
 \item{gam_vb}{Posterior inclusion probability matrix of size p x d.
               Entry (s, t) corresponds to the posterior probability of
               association between predictor s and response t.}
 \item{om_vb}{Vector of length p containing the posterior mean of omega.
              Entry s controls the proportion of responses associated with
              candidate predictor s.}
}
\description{
This function runs \code{\link{locus}} on data with permuted response sample
labels. A common use is for estimatimation of data-driven false discovery
rates of given thresholds on the posterior probabilities of inclusion.
}
\examples{
seed <- 123; set.seed(seed)

###################
## Simulate data ##
###################

## Example using small problem sizes:
##
n <- 200; p <- 250; p0 <- 20; d <- 25; d0 <- 20

## Candidate predictors (subject to selection)
##
# Here we simulate common genetic variants (but any type of candidate
# predictors can be supplied).
# 0 = homozygous, major allele, 1 = heterozygous, 2 = homozygous, minor allele
#
X_act <- matrix(rbinom(n * p0, size = 2, p = 0.25), nrow = n)
X_inact <- matrix(rbinom(n * (p - p0), size = 2, p = 0.25), nrow = n)

shuff_x_ind <- sample(p)
X <- cbind(X_act, X_inact)[, shuff_x_ind]

bool_x_act <- shuff_x_ind <= p0

pat_act <- beta <- matrix(0, nrow = p0, ncol = d0)
pat_act[sample(p0*d0, floor(p0*d0/5))] <- 1
beta[as.logical(pat_act)] <-  rnorm(sum(pat_act))

## Gaussian responses
##
Y_act <- matrix(rnorm(n * d0, mean = X_act \%*\% beta, sd = 0.5), nrow = n)
Y_inact <- matrix(rnorm(n * (d - d0), sd = 0.5), nrow = n)
shuff_y_ind <- sample(d)
Y <- cbind(Y_act, Y_inact)[, shuff_y_ind]

########################
## Infer associations ##
########################

list_perm <- generate_null(n_perm = 10, Y, X, p0_av = p0, link = "identity",
                           user_seed = seed, verbose = FALSE)

}
\seealso{
@seealso \code{\link{locus}}, \code{\link{set_hyper}},
  \code{\link{set_init}}, \code{\link{set_blocks}}, \code{\link{set_groups}}.
}
