% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_null.R
\name{generate_null}
\alias{generate_null}
\title{Generate empirical null distribution using permuted data.}
\usage{
generate_null(n_perm, Y, X, p0_av, Z = NULL, family = "gaussian",
  list_hyper = NULL, list_init = NULL, list_blocks = NULL,
  user_seed = NULL, tol = 0.001, maxit = 1000, batch = TRUE,
  verbose = TRUE, results_dir = NULL, n_cpus = 1)
}
\arguments{
\item{n_perm}{Number of permuted datasets on which \code{\link{locus}} is run.}

\item{Y}{Continuous response data matrix (without permuted sample indices)
of dimension n x d, where n is the number of observations and d is the
number of response variables.}

\item{X}{Input matrix of dimension n x p, where p is the number of candidate
predictors. \code{X} cannot contain NAs. No intercept must be supplied.}

\item{p0_av}{Prior average number of predictors expected to be included in
the model (if \code{list_blocks} is \code{TRUE}, average number per
predictor blocks). Must be \code{NULL} if \code{list_init} and
\code{list_hyper} are both non-\code{NULL} or if \code{list_cv} is
non-\code{NULL}. If \code{list_blocks} is \code{NULL}, can also be a vector
of length p with entry s corresponding to the prior probability that
candidate predictor s is associated with at least one response. If
\code{list_blocks} is non-\code{NULL}, can be a vector of size given by the
number of blocks, with each entry corresponding to the prior average number
of predictors from each block expected to be included in the model.}

\item{Z}{Covariate matrix of dimension n x q, where q is the number of
covariates. \code{NULL} if no covariate. Factor covariates must be supplied
after transformation to dummy coding. No intercept must be supplied.}

\item{family}{Response type. Must be either "\code{gaussian}" for linear
regression or "\code{binomial}" for logistic regression.}

\item{list_hyper}{An object of class "\code{hyper}" containing the model
hyperparameters. Must be filled using the \code{\link{set_hyper}}
function or must be \code{NULL} for default hyperparameters.}

\item{list_init}{An object of class "\code{init}" containing the initial
variational parameters. Must be filled using the
\code{\link{set_init}} function or be \code{NULL} for a default
initialization.}

\item{list_blocks}{An object of class "\code{blocks}" containing setting for
parallel inference on a partitioned predictor space. Must be filled using
the \code{\link{set_blocks}} function or must be \code{NULL} for no
partitioning.}

\item{user_seed}{Seed set for reproducible default choices of hyperparameters
(if \code{list_hyper} is \code{NULL}) and inital variational parameters (if
\code{list_init} is \code{NULL}). Also used at the cross-validation stage
(if \code{list_cv} is non-\code{NULL}). Default is \code{NULL}, no seed set.}

\item{tol}{Tolerance for the stopping criterion.}

\item{maxit}{Maximum number of iterations allowed.}

\item{batch}{If \code{TRUE} a fast batch updating scheme is used (recommended).}

\item{verbose}{If \code{TRUE}, messages are displayed during execution.}

\item{results_dir}{Path where the output of each of the \code{n_perm} runs
will be saved. Default is \code{NULL}, the output is not saved to files and
a list of objects of class "\code{perm}" is returned instead.}

\item{n_cpus}{Number of CPUs to be used. If large, one should ensure that
enough RAM will be available for parallel execution. Set to 1 for serial
execution.}
}
\value{
If \code{results_dir} is \code{NULL}, list of length \code{n_perm}
  with each element corresponding to the output of \code{\link{locus}} on one
  permuted dataset. Each contains:
 \item{ind_perm}{Vector of size n containing the permuted response sample
                 labels.}
 \item{gam_vb}{Posterior inclusion probability matrix of dimension p x d.
               Entry (s, t) corresponds to the posterior probability of
               association between predictor s and response t.}
 \item{om_vb}{Vector of size p containing the posterior mean of omega. Entry
             s controls the proportion of responses associated with predictor
             s.}
}
\description{
This function runs \code{\link{locus}} on data with permuted response sample
labels. A common use is for estimatimation of data-driven false discovery
rates of given thresholds on the posterior probabilities of inclusion.
}
\examples{
user_seed <- 123; set.seed(user_seed)
n <- 200; p <- 250; p0 <- 20; d <- 25; d0 <- 20
list_X <- generate_snps(n = n, p = p)
list_Y <- generate_phenos(n = n, d = d, var_err = 0.25)

dat <- generate_dependence(list_X, list_Y, ind_d0 = sample(1:d, d0),
                           ind_p0 = sample(1:p, p0), vec_prob_sh = 0.05,
                           max_tot_pve = 0.5)

list_perm <- generate_null(n_perm = 10, dat$phenos, dat$snps, p0_av = p0,
                           family = "gaussian", user_seed = user_seed,
                           verbose = FALSE)

}
\seealso{
\code{\link{locus}}
}

