% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_phenos}
\alias{generate_phenos}
\title{Generate phenotypic data with prespecified block-wise correlation structure.}
\usage{
generate_phenos(n, d, var_err, cor_type = NULL, vec_rho = NULL,
  n_cpus = 1)
}
\description{
Generate phenotypic data with prespecified block-wise correlation structure.
}

