% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_snps}
\alias{generate_snps}
\title{Generate SNP data with prespecified spatial correlation structure.}
\usage{
generate_snps(n, p, cor_type = NULL, vec_rho = NULL, vec_maf = NULL,
  n_cpus = 1, user_seed = NULL)
}
\arguments{
\item{n}{Number of observations.}

\item{p}{Number of SNPs.}

\item{cor_type}{String describing the type of dependence structure. The SNPs
can \code{autocorrelated}, \code{equicorrelated}. Set to \code{NULL} for
independent SNPs.}

\item{vec_rho}{Vector of correlation coefficients. Its length determines the
number of blocks of correlated SNPs. Must be smaller than p. Set to
\code{NULL} if independent SNPs.}

\item{vec_maf}{Vector of size p containing the reference minor allele
frequencies used to draw the SNPs. If \code{NULL}, the minor allele
frequencies drawn uniformly at random between 0.05 and 0.5.}

\item{n_cpus}{Number of CPUs used when simulating correlated SNP blocks.
Ignored if independent SNPs. Set to 1 for serial execution.}

\item{user_seed}{Seed set for reproducibility. Default is \code{NULL}, no
seed set.}
}
\value{
An object of class "\code{sim_snps}".
 \item{snps}{Matrix containing the generated SNP data.}
 \item{vec_maf}{Vector containing the SNP sample minor allele frequencies.}
}
\description{
This function generates SNPs under Hardy-Weinberg equilibrium with specific
block correlation structure and minor allele frequencies.
}
\examples{
user_seed <- 123; set.seed(user_seed)
n <- 500; p <- 10000
cor_type <- "autocorrelated"; vec_rho <- runif(100, min = 0.25, max = 0.95)
list_snps <- generate_snps(n, p, cor_type, vec_rho, n_cpus = 2,
                           user_seed = user_seed)

}
\seealso{
\code{\link{replicate_real_snps}}, \code{\link{generate_phenos}},
  \code{\link{replicate_real_phenos}}, \code{\link{generate_dependence}}
}

