% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus.R
\name{locus}
\alias{locus}
\title{Fit a sparse multivariate regression model using variational inference.}
\usage{
locus(Y, X, p_guess, Z = NULL, list_hyper = NULL, list_init = NULL,
  list_cv = NULL, user_seed = NULL, tol = 0.001, maxit = 1000,
  batch = T, save_hyper = F, save_init = F, verbose = T)
}
\arguments{
\item{Y}{continuous response data matrix of dimension n x d, where n is the
number of observations and d is the number of response variables.}

\item{X}{input matrix of dimension n x p, where p is the number of candidate
predictors. X cannot contain NAs.}

\item{p_guess}{is the 'a priori' average number of predictors included in the
model.}

\item{Z}{covariate matrix of dimension n x q, where q is the number of
covariates. NULL if no covariate. Factor covariates must be supplied after
transformation to dummy coding. \code{locus} standardizes the variables so
no intercept must be supplied.}
}
\description{
\code{locus} applies a variational approximation procedure to fit a sparse
multivariate regression model allowing combined selection of predictors and
associated responses in high-dimensional set-ups. Dependence across responses
linked to the same predictors is captured through the model hierarchical
structure.
}

