% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus.R
\name{locus}
\alias{locus}
\title{Fit a sparse multivariate regression model using variational inference.}
\usage{
locus(Y, X, p0_av, Z = NULL, family = "gaussian", list_hyper = NULL,
  list_init = NULL, list_cv = NULL, list_blocks = NULL,
  user_seed = NULL, tol = 1e-04, maxit = 1000, batch = TRUE,
  save_hyper = FALSE, save_init = FALSE, verbose = TRUE)
}
\arguments{
\item{Y}{Continuous response data matrix of dimension n x d, where n is the
number of observations and d is the number of response variables.}

\item{X}{Input matrix of dimension n x p, where p is the number of candidate
predictors. \code{X} cannot contain NAs. No intercept must be supplied.}

\item{p0_av}{Prior average number of predictors expected to be included in
the model. Must be \code{NULL} if \code{list_init} and
\code{list_hyper} are both non-\code{NULL} or if \code{list_cv} is
non-\code{NULL}. Can also be a vector of length p with entry s
corresponding to the prior probability that candidate predictor s is
associated with at least one response.}

\item{Z}{Covariate matrix of dimension n x q, where q is the number of
covariates. \code{NULL} if no covariate. Factor covariates must be supplied
after transformation to dummy coding. No intercept must be supplied.}

\item{family}{Response type. Must be either "\code{gaussian}" for linear
regression or "\code{binomial}" for logistic regression.}

\item{list_hyper}{An object of class "\code{hyper}" containing the model
hyperparameters. Must be filled using the \code{\link{set_hyper}}
function or must be \code{NULL} for default hyperparameters.}

\item{list_init}{An object of class "\code{init}" containing the initial
variational parameters. Must be filled using the
\code{\link{set_init}} function or be \code{NULL} for a default
initialization.}

\item{list_cv}{An object of class "\code{cv}" containing settings for
choosing the prior average number of predictors expected to be included in
the model, \code{p0_av}, by cross-validation. Must be filled using the
\code{\link{set_cv}} function or must be \code{NULL} for no
cross-validation. If non-\code{NULL}, \code{p0_av}, \code{list_init} and
\code{list_hyper} must all be \code{NULL}.}

\item{list_blocks}{An object of class "\code{blocks}" containing setting for
parallel inference on a partitioned predictor space. Must be filled using
the \code{\link{set_blocks}} function or must be \code{NULL} for no
partitioning.}

\item{user_seed}{Seed set for reproducible default choices of hyperparameters
(if \code{list_hyper} is \code{NULL}) and inital variational parameters (if
\code{list_init} is \code{NULL}). Also used at the cross-validation stage
(if \code{list_cv} is non-\code{NULL}). Default is \code{NULL}, no seed set.}

\item{tol}{Tolerance for the stopping criterion.}

\item{maxit}{Maximum number of iterations allowed.}

\item{batch}{If \code{TRUE} a fast batch updating scheme is used
(recommended).}

\item{save_hyper}{If \code{TRUE}, the hyperparameters used for the model are
saved as output.}

\item{save_init}{If \code{TRUE}, the initial variational parameters used for
the inference are saved as output.}

\item{verbose}{If \code{TRUE}, messages are displayed during execution.}
}
\value{
An object of class "\code{vb}" containing the following variational
  estimates and settings:
 \item{lb_opt}{Optimized variational lower bound for the marginal
               log-likelihood.}
 \item{gam_vb}{Posterior inclusion probability matrix of dimension p x d.
               Entry (s, t) corresponds to the posterior probability of
               association between predictor s and response t.}
 \item{mu_alpha_vb}{Matrix of dimension q x d whose entries are the posterior
                    mean regression coefficients for the covariates provided
                    in \code{Z}. \code{NULL} if \code{Z} is \code{NULL}.}
 \item{om_vb}{Vector of size p containing the posterior mean of omega. Entry
             s controls the proportion of responses associated with predictor
             s.}
 \item{x_prpnst}{Vector of size p containing the sums over the colums of
                 \code{gam_vb}, used to evaluate the propensity for the
                 candidate predictors to be involved in associations.}
 \item{y_prpnst}{Vector of size d containing the sums over the rows of
                 \code{gam_vb}, used to evaluate the propensity for the
                 responses to be involved in associations.}
 \item{rmvd_cst_x, rmvd_cst_z}{Vectors containing the indices of constant
                               variables in X (resp. Z) removed prior to the
                               analysis.}
 \item{rmvd_coll_x, rmvd_coll_z}{Vectors containing the indices of variables
                                 in X (resp. Z) removed prior to the
                                 analysis because collinear to other
                                 variables.}
 \item{list_hyper, list_init}{If \code{save_hyper}, resp. \code{save_init},
                              \code{TRUE}, hyperparameters, resp. initial
                              variational parameters, used for inference are
                              saved as output.}
}
\description{
Variational approximation procedure fitting a sparse multivariate regression
model for combined selection of predictors and associated responses in
high-dimensional set-ups. Dependence across responses linked to the same
predictors is captured through the model hierarchical structure.
}
\examples{
user_seed <- 123; set.seed(user_seed)
n <- 200; p <- 300; p0 <- 50; d <- 40; d0 <- 30
list_X <- generate_snps(n = n, p = p)
list_Y <- generate_phenos(n = n, d = d, var_err = 0.25)

dat <- generate_dependence(list_snps = list_X, list_phenos = list_Y,
                           ind_d0 = sample(1:d, d0), ind_p0 = sample(1:p, p0),
                           vec_prob_sh = 0.1, max_tot_pve = 0.9)

vb <- locus(Y = dat$phenos, X = dat$snps, p0_av = p0, family = "gaussian",
            user_seed = user_seed)

}
\seealso{
\code{\link{set_hyper}}, \code{\link{set_init}},
  \code{\link{set_cv}}, \code{\link{set_blocks}}
}

