% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{replicate_real_snps}
\alias{replicate_real_snps}
\title{Generate SNPs emulating real SNP data at hand.}
\usage{
replicate_real_snps(n, real_snps, bl_lgth, p = NULL, maf_thres = NULL,
  n_cpus = 1, user_seed = NULL)
}
\arguments{
\item{n}{Number of observations.}

\item{real_snps}{Matrix of real SNPs (rows observations, columns SNP
variables), without missing values. The entries must be 0, 1 or 2.}

\item{bl_lgth}{Number of variables per block for reproducing the dependence
structure of real SNPs. Must be between 2 and p. Must be small enough
(e.g. 1000) for tractability reasons.}

\item{p}{Number of SNPs. If \code{NULL}, the total number of SNPs
available in the real_snps matrix is used (default).}

\item{maf_thres}{Lower bound for sample minor allele frequencies. Simulated
SNPs with lower sample minor allele frequencies are excluded (as a result
the number of simulated SNPs can be smaller than p). Default is \code{NULL}
for no exclusion.}

\item{n_cpus}{Number of CPUs used when simulating SNP by blocks. Set to 1 for
serial execution.}

\item{user_seed}{Seed set for reproducibility. Default is \code{NULL}, no
seed set.}
}
\value{
An object of class "\code{sim_snps}".
 \item{snps}{Matrix containing the generated SNP data.}
 \item{vec_maf}{Vector containing the SNP sample minor allele frequencies.}
}
\description{
This function simulates SNPs from real SNP data based on their sample minor
allele frequencies and correlation structure.
}
\examples{

user_seed <- 123; set.seed(user_seed)
n <- 500; p <- 7500
cor_type <- "autocorrelated"; vec_rho <- runif(100, min = 0.25, max = 0.95)
list_fake_real_snps <- generate_snps(n, p, cor_type, vec_rho, n_cpus = 2,
                           user_seed = user_seed)
list_snps <- replicate_real_snps(n, list_fake_real_snps$snps, bl_lgth = 100,
                                 n_cpus = 2, user_seed = user_seed)

}
\seealso{
\code{\link{generate_snps}}, \code{\link{generate_phenos}},
  \code{\link{replicate_real_phenos}}, \code{\link{generate_dependence}}
}

